switch back to regular sin and cos

+++ robot_controller/map.cpp
@@ -43,8 +43,8 @@ std::vector<pdd> reportedVictims;
 
 pdd lidarToPoint(const pdd& pos, long double dist, long double absAngle)
 {
-    double x = dist*std::sinl(absAngle) + pos.first;
-    double y = dist*std::cosl(absAngle) + pos.second;
+    double x = dist*std::sin(absAngle) + pos.first;
+    double y = dist*std::cos(absAngle) + pos.second;
 
     return pdd(x, y);
 }
@@ -63,7 +63,7 @@ void update_regions_map(RobotInstance* rb, const float *lidar_image, float theta
         if(std::isinf(dist))
             continue;
 
-        dist *= std::cosl(LIDAR_TILT_ANGLE);
+        dist *= std::cos(LIDAR_TILT_ANGLE);
 
         if(dist > 0.3)
             continue;
